requirements.txt
requests
pandas
load-dotenv
